Moved file reading methods to separate module
