+++ accuread.py
@@ -8,7 +8,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import scipy.io as sio
 from scipy.ndimage.filters import gaussian_filter1d as gaussf
-
+from file_reading import *
 
 class ReadART(object):
     '''Reads the output text files from AccuRT, and includes methods for
@@ -183,207 +183,6 @@ class ReadART(object):
                                'version.txt'), 'r') as ver:
             self.modelversion = ver.readline()[:-1]
 
-    def read_irradiance(self, filename):
-        '''Read output irradiance textfiles from AccuRT model.
-        Returns number of runs, streams, detector depths and wavelengths,
-        and numpy arrays of depths, wavelengths and irradiance'''
-
-        with open(filename, 'r') as infile:
-
-            # read number of runs, streams, depths, wavelengths
-            # and lists of detector depths, wavelengths
-            nruns = int(infile.readline())
-            nstreams = int(infile.readline())
-            ndepths, nwavelengths = [int(j) for j in infile.readline().split()]
-            depths = np.array([float(j) for j in infile.readline().split()])
-            wavelengths = np.array([float(j) for j
-                                    in infile.readline().split()])
-
-            # initiate array for irradiances
-            irradiances = np.empty((ndepths, nwavelengths, nruns))
-
-            # read values for first run
-            for j in range(ndepths):
-                irradiances[j, :, 0] = \
-                    [float(n) for n in infile.readline().split()]
-
-            # read values for rest of runs
-            for i in range(1, nruns):
-                # skip lines with nstreams, ndepths, etc.
-                for _ in range(4):
-                    next(infile)
-                # read values
-                for j in range(ndepths):
-                    irradiances[j, :, i] = \
-                        [float(n) for n in infile.readline().split()]
-
-        return nruns, nstreams, ndepths, nwavelengths, depths,\
-            wavelengths, irradiances
-
-    def read_radiance(self, filename):
-        '''Read output radiance.txt from AccuRT.
-        Returns number of runs, streams, detector depths and wavelengths,
-        and numpy arrays of dephts, wavelengths, polar and azimuth angles
-        and radiance.
-
-        Dimensions of radiance array is
-        (depth) x (wavelength) x (polar angle) x (azimuth angle) x (run number)
-        '''
-
-        with open(filename, 'r') as infile:
-            nruns = int(infile.readline())
-            nstreams = int(infile.readline())
-            ndepths, nwavelengths, npolarangles, nazimuthangles = \
-                [int(j) for j in infile.readline().split()]
-            depths = np.array([float(j) for j in infile.readline().split()])
-            wavelengths = np.array([float(j) for j
-                                    in infile.readline().split()])
-            polarangles = np.array([float(j) for j
-                                    in infile.readline().split()])
-            azimuthangles = np.array([float(j) for j
-                                      in infile.readline().split()])
-
-            radiances = np.empty((ndepths, nwavelengths, npolarangles,
-                                  nazimuthangles, nruns))
-
-            rad = np.array([float(j) for j in infile.readline().split()])
-            radiances[:, :, :, :, 0] = rad.reshape(ndepths,
-                                                   nwavelengths,
-                                                   npolarangles,
-                                                   nazimuthangles)
-
-            for i in range(1, nruns):
-                # skip lines with nstreams, ndepths, etc.
-                for _ in range(6):
-                    next(infile)
-                rad = np.array([float(j) for j in infile.readline().split()])
-                # read values
-                radiances[:, :, :, :, i] = rad.reshape(ndepths,
-                                                       nwavelengths,
-                                                       npolarangles,
-                                                       nazimuthangles)
-
-        return radiances, polarangles, azimuthangles, nruns, nstreams,\
-            ndepths, nwavelengths, depths, wavelengths
-
-    def read_iops(self, filename):
-        '''Read iops.txt, returns dict.'''
-
-        with open(filename, 'r') as infile:
-            nruns = int(infile.readline())
-
-            total_optical_depth = []
-            absorption_coefficients = []
-            scattering_coefficients = []
-            scattering_scaling_factors = []
-            phase_moments = []
-            layer_depths = []
-            wavelengths = []
-
-            for _ in range(nruns):
-                nlayerdepths, nwavelengths, nphasemoments = \
-                    [int(x) for x in infile.readline().split()]
-
-                layer_depths.append(np.array([float(x) for x in
-                                              infile.readline().split()]))
-                wavelengths.append(np.array([float(x) for x in
-                                             infile.readline().split()]))
-
-                _ToD = np.empty((nlayerdepths, nwavelengths))
-                _AC = np.empty((nlayerdepths, nwavelengths))
-                _SC = np.empty((nlayerdepths, nwavelengths))
-                _SSF = np.empty((nlayerdepths, nwavelengths))
-                _PM = np.empty((nlayerdepths, nwavelengths, nphasemoments))
-
-                for j in range(nlayerdepths):
-                    for k in range(nwavelengths):
-                        line = infile.readline().split()
-                        _ToD[j, k] = float(line.pop(0))
-                        _AC[j, k] = float(line.pop(0))
-                        _SC[j, k] = float(line.pop(0))
-                        _SSF[j, k] = float(line.pop(0))
-                        _PM[j, k, :] = np.array(line, dtype='float')
-
-                total_optical_depth.append(_ToD.copy())
-                absorption_coefficients.append(_AC.copy())
-                scattering_coefficients.append(_SC.copy())
-                scattering_scaling_factors.append(_SSF.copy())
-                phase_moments.append(_PM.copy())
-
-            iops = dict(nruns=nruns,
-                        layer_depths=np.squeeze(layer_depths),
-                        wavelengths=np.squeeze(wavelengths),
-                        total_optical_depth=np.squeeze(total_optical_depth),
-                        absorption_coefficients=np.squeeze(
-                            absorption_coefficients),
-                        scattering_coefficients=np.squeeze(
-                            scattering_coefficients),
-                        scattering_scaling_factors=np.squeeze(
-                            scattering_scaling_factors),
-                        phase_moments=np.squeeze(phase_moments))
-
-            return iops
-
-    def read_material_profile(self, filename):
-        '''Read material_profile.txt'''
-        material_profile = []
-        with open(filename) as material_file:
-            while True:
-                line = material_file.readline()
-                if line.startswith('runNo'):
-                    run = int(line.split()[1])
-                    break
-            endoffile = False
-            while True:
-                if endoffile:
-                    break
-                layer = -1
-                material_profile.append([])
-                while True:
-                    line = material_file.readline()
-                    print(line)
-                    if line.startswith('=') or \
-                       line.startswith('-') or \
-                       line.startswith('~'):
-                        pass
-                    elif len(line) == 0:
-                        endoffile = True
-                        break
-                    elif line.startswith('runNo'):
-                        run = int(line.split()[1])
-                        break
-                    elif line.startswith('Layer '):
-                        layer += 1
-                        material_profile[run].append(dict())
-                    elif line.startswith('Bottom depth'):
-                        z = float(line.split()[-2])
-                        material_profile[run][layer]['bottomdepth'] = z
-                    else:
-                        matname = line.replace(' ', '').strip()
-                        material = dict()
-                        conc, conctype = material_file.readline().split()
-                        tau, ssa, g = [float(x) for x in
-                                       material_file.readline().split()]
-                        atau, btau = [float(x) for x in
-                                      material_file.readline().split()]
-                        a, b = [float(x) for x in
-                                material_file.readline().split()]
-                        df = float(material_file.readline())
-
-                        material['concentration'] = float(conc)
-                        material['concentrationtype'] = conctype[1:-1]
-                        material['opticaldepth'] = tau
-                        material['singlescatteringalbedo'] = ssa
-                        material['asymmetryfactor'] = g
-                        material['absorptionoptdep'] = atau
-                        material['scatteringoptdep'] = btau
-                        material['absorption'] = a
-                        material['scattering'] = b
-                        material['deltafitscalingfactor'] = df
-
-                        material_profile[run][layer][matname] = material
-
-        return material_profile
 
     def writefile(self, filename, output=None):
         '''output is 'matlab' or 'netcdf'.'''
